backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
groq>=0.4.1
python-dotenv>=1.0.0
transformers>=4.35.0
torch>=2.2.0
sentencepiece>=0.2.0
accelerate>=0.25.0
pyttsx3>=2.90
SpeechRecognition>=3.10.0
python-multipart>=0.0.6
pydantic>=2.5.0
slowapi>=0.1.9
python-jose>=3.3.0
cachetools>=5.3.0
requests>=2.31.0
